Isolate OpenCL code into a separate folder
