+++ Makefile
@@ -17,7 +17,7 @@ SRC_FILES := $(wildcard src/*.c) $(wildcard src/**/*.c)
 
 nonceMiner: $(SRC_FILES) | bin bin/OpenCL
 	gcc $^ $(CFLAGS) -o bin/$@ $(libs)
-	$(CPF) OpenCL$(PSEP)*.cl bin$(PSEP)OpenCL
+	$(CPF) src$(PSEP)worker$(PSEP)kernel$(PSEP)*.cl bin$(PSEP)OpenCL
 
 nonceMiner_minimal: $(filter-out src/nonceMiner.c, $(SRC_FILES)) test/nonceMiner_minimal.c | bin
 	gcc $^ $(CFLAGS) -o bin/$@ $(libs)
@@ -27,7 +27,7 @@ nonceMiner_minimal_xxhash: $(filter-out src/nonceMiner.c, $(SRC_FILES)) test/non
 
 benchmark: $(filter-out src/nonceMiner.c, $(SRC_FILES)) test/benchmark.c | bin bin/OpenCL
 	gcc $^ $(CFLAGS) -o bin/$@ $(openssl_libs) $(opencl_libs) -lm
-	$(CPF) OpenCL$(PSEP)*.cl bin$(PSEP)OpenCL
+	$(CPF) src$(PSEP)worker$(PSEP)kernel$(PSEP)*.cl bin$(PSEP)OpenCL
 
 bin/OpenCL:
 	mkdir bin$(PSEP)OpenCL
