+++ src/utils/table.h
@@ -1,7 +1,6 @@
 #ifndef TABLE_H
 #define TABLE_H
 
-extern const char three_digit_table[1000*3];
 extern const char four_digit_table[10000*4];
 
 #endif
