+++ src/worker/kernel/buffer_structs_template.cl
